backend/auth_service/app/requirements.txt
python-dotenv
Flask
Flask-SQLAlchemy
psycopg2-binary
bcrypt==4.0.1
PyJWT